pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "django-production"
readme = "README.md"
authors = [{name = "Peter Baumgartner", email = "pete@lincolnloop.com"}]
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Framework :: Django",
]
dynamic = ["version", "description"]
dependencies = [
    "django-upgrade",
    "django-environ",
    "whitenoise",
    "django-webserver[gunicorn]",
    "django-alive",
]
keywords = ["django", "production", "deployment"]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "urllib3",
]

[project.scripts]
django-production-apply = "django_production.__main__:do_patch"

[project.urls]
Home = "https://github.com/lincolnloop/django-production"
Issues = "https://github.com/lincolnloop/django-production/issues"
Changelog = "https://github.com/lincolnloop/django-production/blob/main/CHANGELOG.md"

[tool.flit.module]
name = "django_production"
